+++ lib/view.dart
@@ -49,6 +49,3 @@ export 'package:mvc_application/src/view/uxutils/view.dart';
 
 /// Preferences
 export 'package:prefs/prefs.dart' show Prefs;
-
-/// I10n Translations
-export 'package:i10n_translator/i10n.dart';