+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 2.0.0
  April 21, 2020
- 
+- Removed the packages, dbutils, i10n_translator, firebase_remote_config, android_alarm_manager
+- Removed all dependencies to Firebase
+    
 ## 1.9.0
  April 21, 2020
 - Introduced the mixin HandleError